refactor: DeckTable을 DecksContent에 통합
-  DeckTable을 감싸고 있던 Card 컴포넌트 삭제
